Refactoring to better organize shared dataclasses
